Organize dashboard into separate modules
